// app/common/enum.ts
export enum PACKAGE_STAKING {
  _DIAMOND = 500000,
  _PLATINUM = 100000,
  _GOLD = 50000,
  _SILVER = 10000,
  _BRONZE = 1000,
}