// app/modules/graphql/types/generated.ts
export interface Scalars {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  BigDecimal: number;
  BigInt: number;
  Long: number;
}
// ...
export interface PoolSimpleEarn {
  id: Scalars["ID"];
  nativeTokenAllowed: Scalars["Boolean"];
  stakingToken: Scalars["String"];
  rewardsToken: Scalars["String"];
  products: ProductSimpleEarn[];
}

export interface SubscribedUser {
  id: Scalars["ID"];
  subscriptionId: Scalars["BigInt"];
  product: ProductSimpleEarn;
  account: Scalars["String"];
  subscriptionDate: Scalars["BigInt"];
  valueDate: Scalars["BigInt"];
  interestEndDate: Scalars["BigInt"];
  redeemedDate: Scalars["BigInt"];
  rewardAmount: Scalars["BigDecimal"];
  redeemed: Scalars["Boolean"];
}

export interface ProductSimpleEarn {
  id: Scalars["ID"];
  pool: PoolSimpleEarn;
  lockPeriod: Scalars["String"];
  rewardRate: Scalars["String"];
  rewardRateEarlier: Scalars["String"];
  price: Scalars["BigInt"];
  limit: Scalars["BigInt"];
  limitPerUser: Scalars["BigInt"];
  sold: Scalars["BigInt"];
  redeemed: Scalars["BigInt"];
  paid: Scalars["BigDecimal"];
  subscriptions: SubscribedUser[];
  endDate: Scalars["BigInt"];
  startDate: Scalars["BigInt"];
  roundDate: Scalars["BigInt"];
}